added timestamps to profile and recipe details
