+++ app/(app)/(tabs)/(home)/index.tsx
@@ -43,28 +43,6 @@ export default function Home() {
     refetch();
   }, []);
 
-  const formatDate = (dateString: string) => {
-    const postDate = new Date(dateString);
-    const today = new Date();
-  
-    // Check if the post is from today
-    const isToday = postDate.toDateString() === today.toDateString();
-  
-    // Check if the post is from this year (but not today)
-    const isThisYear = postDate.getFullYear() === today.getFullYear();
-  
-    if (isToday) {
-      // Return time in HH:MM format if today
-      return postDate.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-    } else if (isThisYear) {
-      // Return Month Day if it's this year (but not today)
-      return postDate.toLocaleDateString("en-US", { month: 'long', day: 'numeric' });
-    } else {
-      // Return MM-DD-YY format if it's not from this year
-      return postDate.toLocaleDateString("en-US", { month: '2-digit', day: '2-digit', year: '2-digit' });
-    }
-  };
-
   if (loading) 
     return (
       <View
@@ -139,20 +117,9 @@ export default function Home() {
                   photoUrl={post.photoUrl}
                   mealName={post.name}
                   tastes={post.tastes}
+                  createdAt={post.createdAt}
+                  lastUpdatedAt={post.lastUpdatedAt}
                 />
-
-              <HStack className="flex justify-between w-full">
-                {/* Post time or date */}
-                <Text className="text-sm text-gray-300 mt-2">
-                  {formatDate(post.createdAt)}
-                </Text>
-
-                {post.lastUpdatedAt && (
-                  <Text className="text-sm text-gray-300 mt-2">
-                    Last Updated At: {formatDate(post.lastUpdatedAt)}
-                  </Text>
-                )}
-              </HStack>
             </View>
             ))
           )}
